Finished project -
The game is complete. It now has 5 levels. Each level has a unique word sound challenge, such as 'TH' sounds or 'S' or 'NG' sounds etc. 
If a user is on a level with 'NG' they will be prompted to speak words such as "going".
Their speech is recorded and sent to googles speech to text api for translation. This is aided by contextual hints. If their speech matches the word on screen it is marked as correct. For every incorrect word their final accuracy score is deducted and each incorrect word is added to an array of incorrectwords.  This array of incorrect words is then emailed to the parent or guardian along with an accuracy score report at the end of each level to determine if the user requires speech therapy intervention.